# backend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/admin/companies/{companyId}": {
            "get": {
                "tags": [
                    "admin-company-controller"
                ],
                "operationId": "getCompanyById",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCompanyResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "admin-company-controller"
                ],
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCompanyResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/announcements": {
            "get": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "getAnnouncements",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAnnouncement"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "createAnnouncement",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Announcement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAnnouncement"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/drives/{driveId}/eligibility": {
            "get": {
                "tags": [
                    "admin-eligibility-controller"
                ],
                "operationId": "getEligibilityCriteria",
                "parameters": [
                    {
                        "name": "driveId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEligibilityCriteriaDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-eligibility-controller"
                ],
                "operationId": "setEligibilityCriteria",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "driveId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EligibilityCriteriaDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEligibilityCriteriaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/companies": {
            "get": {
                "tags": [
                    "admin-company-controller"
                ],
                "operationId": "getAllCompanies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListCompanyResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-company-controller"
                ],
                "operationId": "addCompany",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCompanyResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/announcements": {
            "get": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "getAllAnnouncements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "createAnnouncement_1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "admin@campushire.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnouncementRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/announcements/{id}": {
            "delete": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "deleteAnnouncement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/announcements/{id}": {
            "delete": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "deleteAnnouncement_1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CompanyRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "industry": {
                        "type": "string"
                    },
                    "visitHistory": {
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ]
            },
            "ApiResponseCompanyResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/CompanyResponseDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CompanyResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "industry": {
                        "type": "string"
                    },
                    "visitHistory": {
                        "type": "string"
                    },
                    "hiringCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "Department": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
...
            "Announcement": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/User"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "universityRegNo": {
                        "type": "string"
                    },
                    "passwordHash": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "FACULTY",
                            "PLACEMENT_HEAD"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponseAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Announcement"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "EligibilityCriteriaDTO": {
                "type": "object",
                "properties": {
                    "minCgpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "minXMarks": {
                        "type": "number",
                        "format": "double"
                    },
                    "minXiiMarks": {
                        "type": "number",
                        "format": "double"
                    },
                    "maxStandingArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxHistoryOfArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "graduationYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "requiredSkills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "allowedDepartmentIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                }
            },
            "ApiResponseEligibilityCriteriaDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/EligibilityCriteriaDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AnnouncementRequestDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "content",
                    "scope",
                    "title"
                ]
            },
            "AnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "departmentName": {
                        "type": "string"
                    },
                    "createdByEmail": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseAnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseString": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Announcement"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListCompanyResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CompanyResponseDTO"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListAnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnnouncementResponseDTO"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}